💥  (design) move to custom: `Emoji|Toast`, inline emoji `Note`
