+++ packages/design-system/src/components/index.ts
@@ -43,7 +43,6 @@ export {
   DialogTitle,
   DialogDescription,
 } from './Dialog'
-export { Emoji, EmojiParser } from './Emoji'
 export { Flex } from './Flex'
 export { FocusArea } from './FocusArea'
 export { Grid } from './Grid'
@@ -79,15 +78,4 @@ export { TabLink } from './TabLink'
 export { TabsList, Tabs, TabsContent, TabsTrigger } from './Tabs'
 export { Text } from './Text'
 export { TextArea } from './TextArea'
-export {
-  ToastProvider,
-  ToastCenter,
-  ToastCenterItem,
-  ToastWarmer,
-  ToastWarmerItem,
-  Toast,
-  ToastDismiss,
-  useToastDispatchers,
-} from './Toast'
-export type { ToastData, ToastType } from './Toast'
 export { Tooltip } from './Tooltip'
