+++ packages/design-system/src/custom/Emoji/index.ts
