+++ packages/design-system/tsup.config.ts
@@ -22,6 +22,8 @@ import pkg from './package.json'
 const entry = [
   'src/components/Card/Show.tsx',
   'src/components/Hero/HeroImage.tsx',
+  'src/custom/Emoji/index.ts',
+  'src/custom/Toast/index.ts',
   'src/custom/Tooltip/index.ts',
   'src/stitches.config.ts',
   'src/hooks/*',
